+++ Content.Shared/Radio/EntitySystems/SharedHeadsetSystem.cs
@@ -1,17 +1,26 @@
+using System.Linq;
+using Content.Shared.Interaction;
 using Content.Shared.Inventory;
 using Content.Shared.Inventory.Events;
 using Content.Shared.Radio.Components;
+using Content.Shared.Verbs;
+using Robust.Shared.Prototypes;
 
 namespace Content.Shared.Radio.EntitySystems;
 
 public abstract class SharedHeadsetSystem : EntitySystem
 {
+    [Dependency] private readonly SharedInteractionSystem _interaction = default!;
+    [Dependency] private readonly IPrototypeManager _prototype = default!;
+
     public override void Initialize()
     {
         base.Initialize();
         SubscribeLocalEvent<HeadsetComponent, InventoryRelayedEvent<GetDefaultRadioChannelEvent>>(OnGetDefault);
         SubscribeLocalEvent<HeadsetComponent, GotEquippedEvent>(OnGotEquipped);
         SubscribeLocalEvent<HeadsetComponent, GotUnequippedEvent>(OnGotUnequipped);
+
+        SubscribeLocalEvent<HeadsetComponent, GetVerbsEvent<Verb>>(OnGetVerbs);
     }
 
     private void OnGetDefault(EntityUid uid, HeadsetComponent component, InventoryRelayedEvent<GetDefaultRadioChannelEvent> args)
@@ -23,7 +32,7 @@ public abstract class SharedHeadsetSystem : EntitySystem
         }
 
         if (TryComp(uid, out EncryptionKeyHolderComponent? keyHolder))
-            args.Args.Channel ??= keyHolder.DefaultChannel; 
+            args.Args.Channel ??= keyHolder.DefaultChannel;
     }
 
     protected virtual void OnGotEquipped(EntityUid uid, HeadsetComponent component, GotEquippedEvent args)
@@ -34,5 +43,46 @@ public abstract class SharedHeadsetSystem : EntitySystem
     protected virtual void OnGotUnequipped(EntityUid uid, HeadsetComponent component, GotUnequippedEvent args)
     {
         component.IsEquipped = false;
+    }
+
+    private void OnGetVerbs(EntityUid uid, HeadsetComponent component, ref GetVerbsEvent<Verb> args)
+    {
+        if (!args.CanInteract || args.Hands == null)
+            return;
+
+        if (component.IsEquipped && Transform(uid).ParentUid != args.User)
+            return;
+
+        if (!_interaction.InRangeUnobstructed(args.User, args.Target))
+            return;
+
+        if (!TryComp(uid, out EncryptionKeyHolderComponent? keyHolder))
+            return;
+
+        foreach ((var channel, var index) in keyHolder.Channels.Select(static (channel, index) => (channel, index)))
+        {
+            var name = _prototype.Index<RadioChannelPrototype>(channel).LocalizedName;
+
+            var toggled = component.ToggledSoundChannels.Contains(channel);
+
+            args.Verbs.Add(new()
+            {
+                Text = toggled ? $"[bold]{name}" : name,
+                Priority = index,
+                Category = VerbCategory.ToggleHeadsetSound,
+                Act = () => ToggleHeadsetSound((uid, component), channel, !toggled)
+            });
+        }
+    }
+
+    /// <summary>
+    /// Toggles channel on given headset to on or off.
+    /// </summary>
+    public static void ToggleHeadsetSound(Entity<HeadsetComponent> headset, string channel, bool on)
+    {
+        if (on)
+            headset.Comp.ToggledSoundChannels.Add(channel);
+        else
+            headset.Comp.ToggledSoundChannels.Remove(channel);
     }
 }
